extract playlist info
