+++ flow-typed/types/playlist-info.js
@@ -1,0 +1,5 @@
+// @flow
+declare type ProviderPlaylistInfoObject = {
+  playlistId: string,
+  ks?: string
+};
